2025.10.27

+++ src/main/java/com/aimercet/advcontainer/container/handler/IContainerHandler.java
@@ -1,6 +1,6 @@
 package com.aimercet.advcontainer.container.handler;
 
-import com.aimercet.advcontainer.bridge.brlib.HandleSourcePlayer;
+import com.aimercet.advcontainer.container.handler.source.HandleSourcePlayer;
 import com.aimercet.advcontainer.container.ContainerManager;
 import com.aimercet.advcontainer.container.IContainer;
 import com.aimercet.advcontainer.container.ISlot;
@@ -8,10 +8,14 @@ import com.aimercet.advcontainer.container.IStock;
 import com.aimercet.advcontainer.container.slotitem.ISlotItem;
 import com.aimercet.advcontainer.container.handler.source.IHandleSource;
 import com.aimercet.advcontainer.container.source.ISourceLocation;
+import com.aimercet.advcontainer.event.container.ActionStage;
+import com.aimercet.advcontainer.event.container.item.ItemPlaceEvent;
+import com.aimercet.advcontainer.event.container.item.ItemRemoveEvent;
 import com.aimercet.advcontainer.util.Coord;
 import com.aimercet.advcontainer.util.SizeInt;
 import com.aimercet.advcontainer.util.Util;
 import com.aimercet.brlib.log.Logger;
+import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 
@@ -20,6 +24,30 @@ import java.util.List;
 
 public interface IContainerHandler
 {
+
+    /**
+     * 返回source的物品如果转移到target其自身以及占用的格子是否在自身范围内
+     */
+    default boolean isTransferOwnSpace(ItemSource source ,ItemSource target)
+    {
+        if(source==null || target==null)return false;
+        ISlot slot = source.get();
+        ISlotItem item = source.getItem();
+        if(item==null || slot ==null)return false;
+
+        SizeInt sourceSize = getItemSize(source.getItem(), slot.isRotate());
+
+        int maxX = source.coord.x + sourceSize.width -1;
+        int maxY = source.coord.y + sourceSize.height -1;
+
+        for (int x = target.coord.x; x < source.coord.x+sourceSize.width; x++)
+            for (int y = target.coord.y; y < source.coord.y+sourceSize.height; y++)
+                if(x>=source.coord.x && x<=maxX && y>=source.coord.y && y<=maxY)
+                    return true;
+
+        return false;
+    }
+
     default boolean checkHandler(IHandleSource handler , IContainer target)
     {
         if(handler==null ||  target==null)return false;
@@ -130,11 +158,11 @@ public interface IContainerHandler
     /**
      * 空间检查
      */
-    default PlaceResult.Type spaceCheck(IHandleSource handleSource, ISlotItem item, ItemSource coord,boolean rotate)
+    default PlaceResult.Type spaceCheck(IHandleSource handleSource, ISlotItem item, ItemSource coord,boolean rotate , ItemSource exclude)
     {
         if(!checkHandler(handleSource,coord.stock.getContainer()))  return PlaceResult.Type.REFUSE;
 
-        SizeInt size = getItemSize(item,coord,rotate);
+        SizeInt size = getItemSize(item,rotate);
         if(size==null || size.size()<=0)                            return PlaceResult.Type.SIZE_ERROR;
         if(size.size() > coord.stock.getEmpty())                    return PlaceResult.Type.FULL;
 
@@ -144,7 +172,7 @@ public interface IContainerHandler
         int endX = coord.coord.x + size.width,endY = coord.coord.y + size.height;
         if(endX>stockSize.width || endY> stockSize.height)          return PlaceResult.Type.BOUND;
 
-        if(Util.isOccupied(coord.stock, coord.coord,size))          return PlaceResult.Type.NO_SPACE;
+        if(Util.isOccupied(coord.stock, coord.coord,size,exclude))  return PlaceResult.Type.NO_SPACE;
 
         return PlaceResult.Type.SUCCESS;
     }
@@ -152,14 +180,14 @@ public interface IContainerHandler
     /**
      * 综合检查
      */
-    default PlaceResult.Type checkPlace(IHandleSource handleSource, ISlotItem item, ItemSource coord , boolean rotate)
+    default PlaceResult.Type checkPlace(IHandleSource handleSource, ISlotItem item, ItemSource coord , boolean rotate , ItemSource exclude)
     {
         if(!checkHandler(handleSource,coord.stock.getContainer()))return PlaceResult.Type.REFUSE;
 
         PlaceResult.Type preCheck = preCheck(handleSource,item, coord,rotate);
         if(preCheck!= PlaceResult.Type.SUCCESS)                 return preCheck;
 
-        PlaceResult.Type spaceCheck = spaceCheck(handleSource,item, coord,rotate);
+        PlaceResult.Type spaceCheck = spaceCheck(handleSource,item, coord,rotate,exclude);
         if(spaceCheck!= PlaceResult.Type.SUCCESS)               return spaceCheck;
 
         return PlaceResult.Type.SUCCESS;
@@ -168,22 +196,27 @@ public interface IContainerHandler
     /**
      * 获取并处理物品大小
      */
-    default SizeInt getItemSize(ISlotItem item , ItemSource coord , boolean rotate)
+    default SizeInt getItemSize(ISlotItem item , boolean rotate)
     {
         SizeInt size = item.getSize();
         if(rotate)size = size.swap();
         return size;
     }
 
-    default PlaceResult place(IHandleSource handleSource, ISlotItem item, ItemSource coord,boolean rotate){return place(handleSource, item, coord, rotate,false);}
-    default PlaceResult place(IHandleSource handleSource, ISlotItem item, ItemSource coord,boolean rotate , boolean bypass)
+    default PlaceResult place(IHandleSource handleSource, ISlotItem item, ItemSource coord,boolean rotate){return place(handleSource, item, coord, rotate,false,null);}
+    default PlaceResult place(IHandleSource handleSource, ISlotItem item, ItemSource coord,boolean rotate,boolean bypass){return place(handleSource, item, coord, rotate,bypass,null);}
+    default PlaceResult place(IHandleSource handleSource, ISlotItem item, ItemSource coord,boolean rotate , boolean bypass,ItemSource exclude)
     {
         if(!bypass){
-            PlaceResult.Type check = checkPlace(handleSource, item, coord, rotate);
+            PlaceResult.Type check = checkPlace(handleSource, item, coord, rotate , exclude);
             if(check != PlaceResult.Type.SUCCESS) return new  PlaceResult(handleSource,item,check,coord,rotate);
+
+            ItemPlaceEvent preEvent = new ItemPlaceEvent(ActionStage.PRE,new PlaceResult(handleSource, item, null, coord,rotate));
+            Bukkit.getPluginManager().callEvent(preEvent);
+            if(preEvent.isCancelled()) return new PlaceResult(handleSource, item, PlaceResult.Type.REFUSE, coord,rotate);
         }
 
-        SizeInt size =  getItemSize(item,coord,rotate);
+        SizeInt size =  getItemSize(item,rotate);
 
         Util.fillSource(coord.stock,coord.coord,size);
         coord.stock.getSlots()[coord.coord.x][coord.coord.y].setItem(item);
@@ -191,6 +224,10 @@ public interface IContainerHandler
 
         PlaceResult result = new PlaceResult(handleSource, item, PlaceResult.Type.SUCCESS, coord,rotate);
         onPlace(result);
+
+        ItemPlaceEvent postEvent = new ItemPlaceEvent(ActionStage.POST,result);
+        Bukkit.getPluginManager().callEvent(postEvent);
+
         return result;
     }
 
@@ -204,14 +241,22 @@ public interface IContainerHandler
         SizeInt stockSize = coord.stock.getSize();
         if(stockSize==null || stockSize.size()<=0)                  return new RemoveResult(handleSource,item, RemoveResult.Type.STOCK_SIZE,coord);
 
-        SizeInt size = getItemSize(item,coord,coord.get().isRotate());
+        SizeInt size = getItemSize(item,coord.get().isRotate());
         if(size==null || size.size()<=0)                            return new RemoveResult(handleSource,item, RemoveResult.Type.SIZE_ERROR,coord);
+
+        ItemRemoveEvent preEvent = new ItemRemoveEvent(ActionStage.PRE,new RemoveResult(handleSource,item,null,coord));
+        Bukkit.getPluginManager().callEvent(preEvent);
+        if(preEvent.isCancelled()) return new RemoveResult(handleSource,item, RemoveResult.Type.REFUSE,coord);
 
         Util.clearSource(coord.stock,coord.coord,size);
         coord.stock.getSlots()[coord.coord.x][coord.coord.y].setItem(null);
 
         RemoveResult result = new RemoveResult(handleSource, item, RemoveResult.Type.SUCCESS, coord);
         onRemove(result);
+
+        ItemRemoveEvent postEvent = new ItemRemoveEvent(ActionStage.POST,result);
+        Bukkit.getPluginManager().callEvent(postEvent);
+
         return result;
     }
